Update ContactPage.htm
TESTING NOTES 
Line 70 - 79 - Radio Names should maybe be changed to reflect the change in name of their respective Labels, i.e. "radioOption">General Enquiry --> "generalEnquiry">General Enquiry

Line 93 - Phone Number as input type 'Number' has button that increments and decrements value in the field. Maybe it should be input type 'tel' which allows for phone number input and if the browser does not support 'tel', then it will be processed as a string

Line 95 - An input for an appointment ID might be useful if the user is asking a follow-up question (i.e. The person or doctor reading the form can refer to a previous appointment) Code would maybe look something like this
<label for="appointmentID">Appointment ID</label><br>
    <input type="text" id="appointmentID" name="Appointment ID" placeholder="Enter Appointment ID if you are referring to a previous appointment (Optional)" ><br>

Line 98 - A label with today's date seems like a common feature. This website (https://www.arclab.com/en/kb/htmlcss/display-date-time-javascript-php-ssi.html) has instructions on how to do it. This is definitely not essential and is unimportant to the funciton of the form.